+++ internal/controller/patches.go
@@ -180,18 +180,6 @@ func ClusterApiTLSCertManagerPatch(secretName string, issuer *cmmetav1.IssuerRef
 		})
 }
 
-func ClusterCSIDriverSeLinuxPatch(apiVersion *utils.APIVersion) ([]kusttypes.Patch, error) {
-	if apiVersion.Compare(&utils.APIVersion{Major: 1, Minor: 25}) >= 0 {
-		return nil, nil
-	}
-
-	return render(
-		cluster.Resources,
-		"patches/csi-driver-no-selinux.yaml",
-		nil,
-	)
-}
-
 func ClusterApiEndpointPatch(url string) ([]kusttypes.Patch, error) {
 	return render(
 		cluster.Resources,