+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [v2.11.0-rc.1] - 2026-07-24
 
 ### Added
@@ -1235,3 +1237,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v2.10.7]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.10.6...v2.10.7
 [v2.10.8]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.10.7...v2.10.8
 [v2.11.0-rc.1]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.10.8...v2.11.0-rc.1
+[Unreleased]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.11.0-rc.1...HEAD