+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix detection of VolumeGroupSnapshots API.
+
 ## [v2.11.0-rc.1] - 2026-07-24
 
 ### Added