+++ docs/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The minimum supported Kubernetes version is now v1.30.
 
+### Fixed
+
+- Actually merge the host's `lvm.conf` and `lvmlocal.conf` into the LVM configuration used by the satellite, and copy over LVM profiles from the host.
+
 ## [v2.11.0] - 2026-08-04
 
 ### Added